convirtiendo tsx
