--- src/components/chart/index.js
@@ -1,5 +1,0 @@
-// ----------------------------------------------------------------------
-
-export { default } from './chart';
-
-export { default as useChart } from './use-chart';
