--- src/components/color-utils/index.js
@@ -1,2 +1,0 @@
-export { default as ColorPicker } from './color-picker';
-export { default as ColorPreview } from './color-preview';
