--- src/components/svg-color/index.js
@@ -1,1 +1,0 @@
-export { default } from './svg-color';
