+++ src/sections/dashboard/header.tsx
@@ -1,0 +1,3 @@
+export default function Header(){
+  return <h1>Header</h1>
+}
