+++ src/sections/dashboard/main.tsx
@@ -1,0 +1,3 @@
+export default function Main(){
+  return <h1>Main</h1>
+}
