+++ src/sections/dashboard/nav.tsx
@@ -1,0 +1,3 @@
+export default function Nav(){
+  return <h1>Navigator</h1>
+}
