+++ src/sections/dashboard/view/app-view.tsx
@@ -1,0 +1,3 @@
+export default function AppView(){
+  return <h1>Vista de Apps</h1>
+}
