+++ src/sections/login/LoginLayout.tsx
@@ -1,0 +1,135 @@
+import { useState } from "react"
+import { useAuth } from "../../auth/AuthProvider"
+import { Navigate } from "react-router-dom";
+import axios from "axios";
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import Paper from '@mui/material/Paper';
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
+import Typography from '@mui/material/Typography';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
+
+const defaultTheme = createTheme();
+
+export default function Login(){
+
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const auth = useAuth();
+    const [error, setError] = useState("")
+
+    const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+      e.preventDefault();
+      try {
+        const response = await axios.post(import.meta.env.VITE_API_KEY, {
+          email,
+          password,
+        });
+        console.log(response.data);
+        if (response.status === 200) {
+            const jwtToken = response.data.token;
+            localStorage.setItem("jwtToken", jwtToken);
+            auth.login(() => {
+                window.location.href = "/dashboard"
+            })
+        }
+        else{
+          console.log("peticion invalida contraseña incorrecta");
+        }
+      }
+      catch (error) {
+        setError("un error ah ocurrido aqui")
+      }
+    };
+
+    if(auth.isAuthenticated && window.location.pathname === "/"){
+        return <Navigate to ="/dashboard" replace />;
+    }
+
+    const handleLogout = () => {
+        localStorage.removeItem("jwtToken");
+        auth.logout();
+      };  
+    
+    return (
+    <>
+        <ThemeProvider theme={defaultTheme}>
+            <Grid container component="main" sx={{ height: '100vh' }}>
+            <CssBaseline />
+                <Grid
+                    item
+                    xs={false}
+                    sm={4}
+                    md={7}
+                    sx={{
+                        backgroundImage: '',
+                        backgroundRepeat: 'no-repeat',
+                        backgroundColor: (t) =>
+                        t.palette.mode === 'light' ? t.palette.grey[50] : t.palette.grey[900],
+                        backgroundSize: 'cover',
+                        backgroundPosition: 'center',
+                    }}
+                />
+                <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
+                <Box
+                    sx={{
+                    my: 8,
+                    mx: 4,
+                    display: 'flex',
+                    flexDirection: 'column',
+                    alignItems: 'center',
+                    }}
+                    
+                >
+                    <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+                        <LockOutlinedIcon />
+                    </Avatar>
+                    <Typography component="h1" variant="h5">
+                        Login Dar
+                    </Typography>
+                    <Box component="form" sx={{ mt: 1 }} onSubmit={handleSubmit}>
+                        <TextField
+                            margin="normal"
+                            required
+                            fullWidth
+                            id="email"
+                            label="Usuario"
+                            name="email"
+                            autoComplete="email"
+                            autoFocus
+                            value={email}
+                            onChange={(e) => setEmail(e.target.value)}
+                        />
+                        <TextField
+                            margin="normal"
+                            required
+                            fullWidth
+                            name="password"
+                            label="Contraseña"
+                            type="password"
+                            id="password"
+                            autoComplete="current-password"
+                            value={password}
+                            onChange={(e) => setPassword(e.target.value)}
+                        />
+                    
+                        <Button
+                            type="submit"
+                            fullWidth
+                            variant="contained"
+                            sx={{ mt: 3, mb: 2 }}
+                        >
+                            Ingresar
+                        </Button>
+                    </Box>
+                  </Box>
+                </Grid>
+            </Grid>
+        </ThemeProvider> 
+    </>
+    )
+}
